Dev front

+++ src/components/MainLayout/MainLayOut.tsx
@@ -19,6 +19,7 @@ import { auth } from '../../store/redux/users/slice'
 import { useDispatch } from 'react-redux'
 import { useLazyLogoutQuery } from '../../api/userLoginService'
 import { RoleE } from 'enum/roleE'
+import { BackgroundAnimation } from '../BackgroundAnimation'
 
 export const MainLayOut: FC = memo(() => {
   const isLogin = useAppSelector(authSelector)
@@ -38,6 +39,7 @@ export const MainLayOut: FC = memo(() => {
   const [currentTariff, setCurrentTariff] = useState<any | null>(null)
   const [showChat, setShowChat] = useState<boolean>(false)
   const [overaiLockExists, setOveraiLockExists] = useState(false)
+  const routesWithoutPrevious = [`/school/${schoolName}/meetings/`]
 
   useEffect(() => {
     if (userRole === 1) {
@@ -80,32 +82,22 @@ export const MainLayOut: FC = memo(() => {
   }, [data])
 
   useEffect(() => {
-    getGroups(schoolName)
-    if (groupsError && 'originalStatus' in groupsError && groupsError.originalStatus === 404) {
-      localStorage.clear()
-      logout()
-      dispatch(auth(false))
-      navigate(generatePath(Path.InitialPage))
+    if (groupsError) {
+      getGroups(schoolName)
+      if (groupsError && 'originalStatus' in groupsError && groupsError.originalStatus === 404) {
+        localStorage.clear()
+        logout()
+        dispatch(auth(false))
+        navigate(generatePath(Path.InitialPage))
+      }
     }
   }, [groupsError, navigate])
 
   return (
     <>
       <div className={styles.wrapper}>
-        <div className={styles.bg}>
-          <div className={styles.bg_wrap1}></div>
-        </div>
-        <div className={styles.bg}>
-          <div className={styles.bg_wrap2}></div>
-        </div>
-        <div className={styles.bg}>
-          <div className={styles.bg_wrap3}></div>
-        </div>
-        <div className={styles.bg}>
-          <div className={styles.bg_wrap4}></div>
-        </div>
+        <BackgroundAnimation />
         {userRole === RoleE.Admin && progress.completion_percentage < 100 && <NewSchoolProgress />}
-        <Navbar />
         <Header />
         <motion.main
           className={styles.container}
@@ -123,7 +115,8 @@ export const MainLayOut: FC = memo(() => {
             duration: 1.2,
           }}
         >
-          <Previous />
+          <Navbar />
+          {!routesWithoutPrevious.includes(location.pathname) && <Previous />}
           <Outlet />
         </motion.main>
         {showChat && isSuccess && <ChatGPT openChatModal={handlers.onToggle} closeChatModal={handlers.off} />}
